:building_construction: Use `groupInfo.deployable` to determine staging env
